Safari-611.1.10/LayoutTests/loader/navigation-policy/should-open-external-urls/api-false-followed-by-api-true-expected.txt

 - decidePolicyForNavigationAction
<NSURLRequest URL resources/do-nothing, main document URL resources/do-nothing, http method GET> is main frame - yes should open URLs externally - no
 - decidePolicyForNavigationAction
<NSURLRequest URL resources/notify-done.html, main document URL resources/notify-done.html, http method GET> is main frame - yes should open URLs externally - yes
